requirements.txt
python-dotenv
openai
requests
pypdf
gradio 